moved functions to tdvp_utils
